create bokeh.testing.plugins
