+++ conftest.py
@@ -1,1 +1,1 @@
-pytest_plugins = "tests.plugins.implicit_mark"
+pytest_plugins = "bokeh.testing.plugins.implicit_mark"
